// app/src/auth/ability/ability.types.ts
export type AppAction =
  | 'manage'
  | 'create'
  | 'read'
  | 'update'
  | 'delete'
  | 'pay'
  | 'send'
  | 'import'
  | 'export'
  | 'report';

export type AppSubject =
  | 'all'
  | 'Audit'
  | 'AccountPlan'
  | 'Alert'
  | 'Calendar'
  | 'Charge'
  | 'ChargeTemplate'
  | 'Customer'
  | 'CustomerPortal'
  | 'Document'
  | 'Finance'
  | 'FinancialEntry'
  | 'Invitation'
  | 'Lead'
  | 'Loan'
  | 'NFe'
  | 'Notification'
  | 'OpenFinance'
  | 'Payable'
  | 'PersonalFinance'
  | 'Product'
  | 'Purchase'
  | 'Reconciliation'
  | 'Sale'
  | 'Search'
  | 'Settings'
  | 'Stock'
  | 'Supplier'
  | 'Task'
  | 'Tax'
  | 'User'
  | 'WhatsappAdmin';